+++ library/common/src/main/java/com/google/android/exoplayer2/Tracks.java
@@ -293,6 +293,7 @@ public final class Tracks implements Bundleable {
    * Returns true if at least one track of type {@code trackType} is {@link
    * Group#isTrackSupported(int, boolean) supported}.
    *
+   * @param trackType The track type to query support for.
    * @param allowExceedsCapabilities Whether to consider the track as supported if it has a
    *     supported {@link Format#sampleMimeType MIME type}, but otherwise exceeds the advertised
    *     capabilities of the device. For example, a video track for which there's a corresponding