+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/FakeTimeline.java
@@ -391,6 +391,7 @@ public final class FakeTimeline extends Timeline {
   /**
    * Creates a fake timeline with the given window definitions.
    *
+   * @param manifests The manifests of the windows.
    * @param windowDefinitions A list of {@link TimelineWindowDefinition}s.
    */
   public FakeTimeline(Object[] manifests, TimelineWindowDefinition... windowDefinitions) {
@@ -401,6 +402,8 @@ public final class FakeTimeline extends Timeline {
    * Creates a fake timeline with the given window definitions and {@link
    * com.google.android.exoplayer2.source.ShuffleOrder}.
    *
+   * @param manifests The manifests of the windows.
+   * @param shuffleOrder A shuffle ordering for the windows.
    * @param windowDefinitions A list of {@link TimelineWindowDefinition}s.
    */
   public FakeTimeline(